+++ test/td_dd/data_structures/data_structures_test.exs
@@ -24,7 +24,7 @@ defmodule TdDd.DataStructuresTest do
     domain = CacheHelpers.insert_domain()
     %{id: template_id, name: template_name} = template = CacheHelpers.insert_template()
 
-    CacheHelpers.insert_structure_type(structure_type: template_name, template_id: template_id)
+    CacheHelpers.insert_structure_type(name: template_name, template_id: template_id)
 
     %{id: system_id} = system = insert(:system, external_id: "test_system")
 
@@ -163,6 +163,31 @@ defmodule TdDd.DataStructuresTest do
       assert %{__meta__: %{state: :deleted}} = data_structure
       assert DataStructures.get_data_structure!(ds2.id) <~> ds2
       assert DataStructures.get_data_structure!(ds3.id) <~> ds3
+    end
+  end
+
+  describe "logical_delete_data_structure/2" do
+    test "logical_delete_data_structure/2 delete the logical data_structure version", %{
+      data_structure_version: %{id: parent_version_id} = data_structure_version,
+      data_structure: %{id: id},
+      claims: claims
+    } do
+      insert(:structure_metadata, data_structure_id: id, version: parent_version_id)
+
+      {:ok, result} = DataStructures.logical_delete_data_structure(data_structure_version, claims)
+
+      assert %{delete_dsv_descendents: {1, nil}} = result
+      assert %{delete_metadata_descendents: {1, nil}} = result
+      assert %{data_structure_version_descendents: [^parent_version_id]} = result.descendents
+      assert %{data_structures_ids: [^id]} = result.descendents
+    end
+
+    test "emits an audit event", %{data_structure_version: data_structure_version, claims: claims} do
+      assert {:ok, %{audit: [event_id | _]}} =
+               DataStructures.logical_delete_data_structure(data_structure_version, claims)
+
+      assert {:ok, [%{id: ^event_id}]} =
+               Stream.range(:redix, @stream, event_id, event_id, transform: :range)
     end
   end
 
@@ -362,6 +387,24 @@ defmodule TdDd.DataStructuresTest do
     end
   end
 
+  describe "template_name/1" do
+    test "returns an empty string if no template or type exists" do
+      %{name: type} = insert(:data_structure_type)
+      dsv = insert(:data_structure_version, type: type)
+      assert DataStructures.template_name(dsv) == ""
+
+      dsv = insert(:data_structure_version)
+      assert DataStructures.template_name(dsv) == ""
+    end
+
+    test "returns the name of the corresponding template" do
+      %{id: template_id, name: name} = CacheHelpers.insert_template()
+      %{name: type} = insert(:data_structure_type, template_id: template_id)
+      dsv = insert(:data_structure_version, type: type)
+      assert DataStructures.template_name(dsv) == name
+    end
+  end
+
   describe "data_structures" do
     @update_attrs %{
       # description: "some updated description",
@@ -452,24 +495,6 @@ defmodule TdDd.DataStructuresTest do
       refute %{"domain_external_id" => "baz", "ou" => "baz"}
              |> put_domain_id(ids)
              |> Map.has_key?("domain_id")
-    end
-
-    test "get_structures_metadata_fields/1 will retrieve all metada fields of the filtered structures" do
-      insert(:data_structure_version, type: "foo", metadata: %{"foo" => "value"})
-      insert(:data_structure_version, type: "foo", metadata: %{"Foo" => "value"})
-      insert(:data_structure_version, type: "foo", metadata: %{"bar" => "value"})
-      insert(:data_structure_version, type: "bar", metadata: %{"xyz" => "value"})
-
-      insert(:data_structure_version,
-        type: "bar",
-        metadata: %{"baz" => "value"},
-        deleted_at: DateTime.utc_now()
-      )
-
-      assert [_ | _] =
-               fields = DataStructures.get_structures_metadata_fields(%{type: ["foo", "bar"]})
-
-      assert Enum.all?(["xyz", "Foo", "bar", "foo"], &(&1 in fields))
     end
   end