+++ test/td_dd/data_structures/structure_notes_workflow_test.exs
@@ -23,7 +23,7 @@ defmodule TdDd.DataStructures.StructureNoteWorkflowTest do
       template = build(:template, name: @template_name, content: content)
 
     {:ok, _} = TemplateCache.put(template, publish: false)
-    CacheHelpers.insert_structure_type(structure_type: template_name, template_id: template_id)
+    CacheHelpers.insert_structure_type(name: template_name, template_id: template_id)
     on_exit(fn -> TemplateCache.delete(template_id) end)
 
     start_supervised!(TdDd.Search.StructureEnricher)